VCTClasses

This file contains the class definitions for the hierarchical structure connecting simulations to trials.

pcvct.AbstractSamplingFolders — Type

AbstractSamplingFolders

A structure representing the folders used for sampling in a VCT (Virtual Cell Technology) environment.

Fields

config_folder::String: Name of the configuration folder.

rulesets_collection_folder::String: Name of the rulesets collection folder.

ic_cell_folder::String: Name of the initial condition (IC) cells folder.

ic_substrate_folder::String: Name of the initial condition (IC) substrate folder.

ic_ecm_folder::String: Name of the initial condition (IC) extracellular matrix (ECM) folder.

custom_code_folder::String: Name of the custom code folder.

pcvct.AbstractSamplingIDs — Type

struct AbstractSamplingIDs

A struct representing various IDs used for sampling in the VCT environment.

Fields

config_id::Int: Integer identifying the base configuration folder ID for lookup in the database.

rulesets_collection_id::Int: Integer identifying which rulesets collection to use as a framework.

ic_cell_id::Int: Integer identifying the initial condition cells folder for lookup in the database.

ic_substrate_id::Int: Integer identifying the initial condition substrate folder for lookup in the database.

ic_ecm_id::Int: Integer identifying the initial condition extracellular matrix folder for lookup in the database.

custom_code_id::Int: Integer identifying the custom code folder (with {main.cpp, Makefile, custom_modules/{custom.cpp, custom.h}} as folder structure) for lookup in the database.