VCTCompilation

This file contains the functions used to compile a PhysiCell project, including using the necessary compiler macros.

pcvct.getCompilerFlags — Method

getCompilerFlags(S::AbstractSampling) -> Tuple{String, Bool, Bool}

Generate the compiler flags for the given sampling object S.

Arguments

S::AbstractSampling: The sampling object for which to generate compiler flags.

Returns

cflags::String: The compiler flags as a string.

recompile::Bool: A boolean indicating whether recompilation is needed.

clean::Bool: A boolean indicating whether cleaning is needed.

Description

This function generates the necessary compiler flags based on the system and the macros defined in the sampling object S. It checks if the system is macOS and adjusts the flags accordingly. It also compares the current macros with the updated macros to determine if recompilation and cleaning are needed.

Notes

On macOS, it checks if the system architecture is arm64 and adjusts the -mfpmath flag accordingly.

It reads the current macros from a file and compares them with the updated macros to decide if recompilation and cleaning are necessary.

If the project file does not exist in the specified directory, it sets recompile to true.

pcvct.loadCustomCode — Method

loadCustomCode(S::AbstractSampling; force_recompile::Bool=false)

Load and compile custom code for a given sampling instance.

Arguments

S::AbstractSampling: An instance of AbstractSampling containing the necessary folder names and IDs for custom code compilation.

force_recompile::Bool=false: A boolean flag to force recompilation of the custom code regardless of the current state.

Description

This function performs the following steps:

Retrieves compiler flags and determines if recompilation is necessary.

If recompilation is not required and force_recompile is false, the function returns immediately.

If recompilation is required, it optionally cleans the build directory.

Copies custom module files, main.cpp, and Makefile from the source directory to the target directory.

Compiles the custom code using the make command with appropriate flags.

Logs the compilation output and errors to specified files.

Deletes the error log file if it is empty.

Cleans up by removing copied files and restoring the default Makefile.

Moves the compiled project to the designated folder.

Notes

The function assumes the presence of specific directory structures and file names.

On macOS, the -j flag is used to parallelize the compilation process.

The function logs compilation output and errors to files in the source directory.