VCTComponents

Allows for combining PhysiCell input components into whole inputs.

Currently, only supports this for intracellular ODE (libRoadRunner) models.

pcvct.PhysiCellComponent — Type

PhysiCellComponent

A struct to hold the information about a component that is used to assemble an input of PhysiCell.

The type and name are the only fields that are compared for equality. The type represents the type of component that it is. Currently, only "roadrunner" is supported. The name is the name of the file inside the components/type/ directory. The path_from_components is the path from the components directory to the file. The id is the id of the component, which will be -1 to indicate it is not yet set. The id is used to link which cell definition(s) use which component(s).