VCTConfiguration

Interface with the configuration file necessary for PhysiCell simulations.

Provide functionality for accessing and modifying elements in any XML, including the PhysiCell configuration file, XML rules file, combined intracellular XML file, XML IC cell file, and XML IC ECM file.

pcvct.createXMLFile — Method

createXMLFile(location::Symbol, M::AbstractMonad)

Create XML file for the given location and variation_id in the given monad.

The file is placed in $(location)_variations and can be accessed from there to run the simulation(s).

pcvct.icCellsPath — Method

icCellsPath(cell_definition::AbstractString, patch_type::AbstractString, patch_id path_elements::Vararg{<:AbstractString})

Return the XML path to the IC cell patch for the given cell definition, patch type, and patch ID. Optionally, add more path_elements to the path as extra arguments.

Example

julia> icCellsPath("default", "disc", 1)
3-element Vector{String}:
 "cell_patches:name:default"
 "patch_collection:type:disc"
 "patch:ID:1"

julia> icCellsPath("default", "disc", 1, "x0")
4-element Vector{String}:
 "cell_patches:name:default"
 "patch_collection:type:disc"
 "patch:ID:1"
 "x0"

pcvct.icECMPath — Method

icECMPath(layer_id::Int, patch_type::AbstractString, patch_id, path_elements::Vararg{<:AbstractString})

Return the XML path to the IC ECM patch for the given layerid, patchtype, and patchid. Optionally, add more pathelements to the path as extra arguments.

Example

julia> icECMPath(2, "ellipse", 1)
3-element Vector{String}:
 "layer:ID:2"
 "patch_collection:type:ellipse"
 "patch:ID:1"

julia> icECMPath(2, "elliptical_disc", 1, "density")
4-element Vector{String}:
 "layer:ID:2"
 "patch_collection:type:elliptical_disc"
 "patch:ID:1"
 "density"

pcvct.rulePath — Method

rulePath(cell_definition::AbstractString, behavior::AbstractString, path_elements::Vararg{<:AbstractString})

Return the XML path to the rule for the given cell definition and behavior.

Optionally, add more path_elements to the path as extra arguments.

Example

julia> rulePath("T cell", "attack rate")
2-element Vector{String}:
 "behavior_ruleset:name:T cell"
 "behavior:name:attack rate"

julia> rulePath("cancer", "cycle entry", "increasing_signals", "signal:name:oxygen", "half_max")
5-element Vector{String}:
 "behavior_ruleset:name:cancer"
 "behavior:name:cycle entry"
 "increasing_signals"
 "signal:name:oxygen"
 "half_max"