VCTDeletion

Safely delete output from a pcvct project.

pcvct.deleteSimulationsByStatus — Function

deleteSimulationsByStatus(status_codes_to_delete::Vector{String}=["Failed"]; user_check::Bool=true)

Delete simulations from the database based on their status codes.

The list of possible status codes is: "Not Started", "Queued", "Running", "Completed", "Failed".

Arguments

status_codes_to_delete::Vector{String}: A vector of status codes for which simulations should be deleted. Default is ["Failed"].

user_check::Bool: If true, prompts the user for confirmation before deleting simulations. Default is true.

pcvct.eraseSimulationID — Method

eraseSimulationID(simulation_id::Int[; monad_id::Union{Missing,Int}=missing])

Erase a simulation ID from the monad it belongs to simulations.csv.

If monad_id is not provided, the function will infer it from the simulation ID. If the monad contains only the given simulation ID, the monad will be deleted.