VCTLoader

Load PhysiCell data into useful forms for downstream analysis.

pcvct.PhysiCellSequence — Type

PhysiCellSequence

A sequence of PhysiCell snapshots.

Fields

folder::String: The folder containing the PhysiCell simulation output.

snapshots::Vector{PhysiCellSnapshot}: A vector of PhysiCell snapshots.

cell_type_to_name_dict::Dict{Int, String}: A dictionary mapping cell type IDs to cell type names.

substrate_names::Vector{String}: A vector of substrate names.

pcvct.PhysiCellSnapshot — Type

PhysiCellSnapshot

A single snapshot of a PhysiCell simulation.

The cells, substrates, and mesh fields may remain empty until they are needed for analysis.

Fields

folder::String: The folder containing the PhysiCell simulation output.

index::Union{Int, Symbol}: The index of the snapshot. Can be an integer or a symbol (:initial or :final).

time::Float64: The time of the snapshot.

cells::DataFrame: A DataFrame containing cell data.

substrates::DataFrame: A DataFrame containing substrate data.

mesh::Dict{String, Vector{Float64}}: A dictionary containing mesh data.

pcvct.computeMeanSpeed — Method

computeMeanSpeed(simulation_id::Integer[; direction=:any])

Return dictionaries containing the mean speed, total distance traveled, and total time spent for each cell type in the PhysiCell simulation.

The time is counted from when the cell first appears in simulation output until it dies or the simulation ends, whichever comes first.

To account for cells that may change cell type during the simulation, the dictionaries returned are keyed by cell type. So, a dictionary with key "A" and value 2.0 indicates that the mean speed of this cell while it was of type "A" is 2.0.

Arguments

simulation_id::Integer: The ID of the PhysiCell simulation.

direction::Symbol: The direction to compute the mean speed. Can be :x, :y, :z, or :any (default). If :x, for example, the mean speed is calculated using only the x component of the cell's movement.

Returns

mean_speed_dicts::Vector{Dict{String,Float64}}: A vector of dictionaries where each dictionary is specific to a single cell. The key is the cell type and the value is the mean speed of that cell.

distance_dicts::Vector{Dict{String,Float64}}: A vector of dictionaries where each dictionary is specific to a single cell. The key is the cell type and the value is the total distance traveled by that cell.

time_dicts::Vector{Dict{String,Float64}}: A vector of dictionaries where each dictionary is specific to a single cell. The key is the cell type and the value is the total time in the simulation for that cell.

pcvct.getCellDataSequence — Method

getCellDataSequence(sequence::PhysiCellSequence, labels::Vector{String}; include_dead::Bool=false, include_cell_type::Bool=false)

Return a dictionary where the keys are cell IDs from the PhysiCell simulation and the values are NamedTuples containing the time and the values of the specified labels for that cell. In the case of a label that has multiple columns, such as position, the values are concatenated into a length(snapshots) x number of columns array.

Examples

data = getCellDataSequence(sequence, ["position", "elapsed_time_in_phase"]; include_dead=true, include_cell_type=true)
data[1] # the first cell's data
data[1].position # an Nx3 array of the cell's position over time
data[1].elapsed_time_in_phase # an Nx1 array of the cell's elapsed time in phase over time
data[1].cell_type_name # the cell type name of the first cell

pcvct.getCellPositionSequence — Method

getCellPositionSequence(sequence::PhysiCellSequence; include_dead::Bool=false, include_cell_type::Bool=false)

Return a dictionary where the keys are cell IDs from the PhysiCell simulation and the values are NamedTuples containing the time and the position of the cell.

This is a convenience function for getCellDataSequence(sequence, "position"; include_dead=include_dead, include_cell_type=include_cell_type).