VCTModule

This file holds some core functionality for pcvct.

pcvct.constituentsType — Method

constituentsType(T::AbstractTrial) -> Type

Returns the type of constituents for a given AbstractTrial.

Arguments

T::AbstractTrial: An AbstractTrial object.

Returns

Type: The type of constituents.

pcvct.initializeVCT — Method

initializeVCT(path_to_physicell::String, path_to_data::String)

Initializes the VCT environment by setting the paths to PhysiCell and data directories, and initializing the database.

Arguments

path_to_physicell::String: Path to the PhysiCell directory.

path_to_data::String: Path to the data directory.

pcvct.pcvctLogo — Method

pcvctLogo() -> String

Returns a string representation of the PCVCT logo.

pcvct.readConstituentIDs — Method

readConstituentIDs(path_to_csv::String) -> Vector{Int}

Reads constituent IDs from a CSV file.

Arguments

path_to_csv::String: Path to the CSV file.

Returns

Vector{Int}: A vector of constituent IDs.

pcvct.readConstituentIDs — Method

readConstituentIDs(T::AbstractTrial)

Reads the constituent IDs for a given trial type T.

Arguments

T::AbstractTrial: An instance of a trial type.

Returns

A list of constituent IDs read from a CSV file.

Details

The function constructs a file path based on the type and ID of the trial T. It then reads the constituent IDs from a CSV file located at the constructed path.