VCTPhysiCellStudio

Launch PhysiCell Studio for a simulation run in pcvct.

pcvct.runStudio — Method

runStudio(simulation_id::Int; python_path::Union{Missing,String}=path_to_python, studio_path::Union{Missing,String}=path_to_studio)

Launch PhysiCell Studio for a given simulation.

Creates temporary config and rules files to avoid overwriting the original files in the output folder. The intent of this function is to allow users to visualize the results of a simulation with Studio, rather than to modify the simulation itself.

The path to the python executable and the Studio folder must be set. pcvct will look for these in the environment variables PCVCT_PYTHON_PATH and PCVCT_STUDIO_PATH, respectively.