VCTRunner

Run simulations, monads, samplings, and trials in the pcvct framework.

Base.run — Method

run(T::AbstractTrial[; force_recompile::Bool=false, prune_options::PruneOptions=PruneOptions()])`

Run the given simulation, monad, sampling, or trial.

Call the appropriate functions to run the simulations and return the number of successful simulations. Also print out messages to the console to inform the user about the progress and results of the simulations.

Arguments

T::AbstractTrial: The trial, sampling, monad, or simulation to run.

force_recompile::Bool=false: If true, forces a recompilation of all files by removing all .o files in the PhysiCell directory.

prune_options::PruneOptions=PruneOptions(): Options for pruning simulations.

pcvct.collectSimulationTasks — Method

collectSimulationTasks(T::AbstractTrial[; force_recompile::Bool=false, prune_options::PruneOptions=PruneOptions()])

Collect the simulation tasks for the given trial, sampling, monad, or simulation.

Used by run to collect the tasks to run.

See also run.

pcvct.runAbstractTrial — Method

runAbstractTrial(T::AbstractTrial; force_recompile::Bool=false, prune_options::PruneOptions=PruneOptions())

Alias for run, but only with this particular signature. Does not work on Cmd objects as Base.run is built for. Also, does not work with running sensitivity samplings.