Deletion

Safely delete output from a pcvct project.

pcvct.deleteAllSimulations — Method

deleteAllSimulations(; delete_supers::Bool=true, and_constraints::String="")

Delete all simulations. See deleteSimulations for the meaning of the keyword arguments.

pcvct.deleteMonad — Method

deleteMonad(monad_ids::AbstractVector{<:Integer}; delete_subs::Bool=true, delete_supers::Bool=true)

Delete monads from the database by monad ID.

Arguments

monad_ids: list of monad IDs to delete. Can also pass an Integer.

delete_subs: If delete_subs is true, will further delete the simulations corresponding to the monad.

delete_supers: If delete_supers is true, will further delete any samplings that no longer have associated monads after the deletion.

pcvct.deleteSampling — Method

deleteSampling(sampling_ids::AbstractVector{<:Integer}; delete_subs::Bool=true, delete_supers::Bool=true)

Delete samplings from the database by sampling ID.

Arguments

sampling_ids: list of sampling IDs to delete. Can also pass an Integer.

delete_subs: If delete_subs is true, will further delete the monads corresponding to the sampling.

delete_supers: If delete_supers is true, will further delete any trials that no longer have associated samplings after the deletion.

pcvct.deleteSimulation — Function

deleteSimulation(args...; kwargs...)

Alias for deleteSimulations.

pcvct.deleteSimulations — Method

deleteSimulations(simulation_ids::AbstractVector{<:Union{Integer,Missing}}; delete_supers::Bool=true, and_constraints::String="")

Deletes the simulations with the input IDs from the database and from the data/outputs/simulations folder.

Works with a vector of simulation IDs, a single simulation ID, a vector of Simulations or a single Simulation. If delete_supers is true, it will also delete any monads, samplings, and trials that no longer have any simulations associated with them. It is recommended to leave this to true to keep the database clean. The and_constraints argument allows for additional SQLite conditions to be added to the WHERE clause of the SQLite query. Use this only after inspecting the simulations table in the data/vct.db database. Note: deleteSimulation is an alias for deleteSimulations.

Examples

deleteSimulations(1:3)
deleteSimulations(4)
deleteSimulations(1:100; and_constraints="AND config_id = 1") # delete simulations with IDs 1 to 100 that have config_id = 1

pcvct.deleteSimulationsByStatus — Function

deleteSimulationsByStatus(status_codes_to_delete::Vector{String}=["Failed"]; user_check::Bool=true)

Delete simulations from the database based on their status codes.

The list of possible status codes is: "Not Started", "Queued", "Running", "Completed", "Failed".

Arguments

status_codes_to_delete::Vector{String}: A vector of status codes for which simulations should be deleted. Default is ["Failed"].

user_check::Bool: If true, prompts the user for confirmation before deleting simulations. Default is true.

pcvct.deleteTrial — Method

deleteTrial(trial_ids::AbstractVector{<:Integer}; delete_subs::Bool=true)

Delete trials from the database by trial ID.

Arguments

trial_ids: list of trial IDs to delete. Can also pass an Integer.

delete_subs: If delete_subs is true, will further delete the samplings corresponding to the trials.

pcvct.eraseSimulationIDFromConstituents — Method

eraseSimulationIDFromConstituents(simulation_id::Int[; monad_id::Union{Missing,Int}=missing])

Erase a simulation ID from the simulations.csv file of the monad it belongs to.

If monad_id is not provided, the function will infer it from the simulation ID. If the monad contains only the given simulation ID, the monad will be deleted. This is used when running simulations if they error so that the monads no longer rely on them, but the simulation output can still be checked.

pcvct.resetDatabase — Method

resetDatabase()

Reset the database (after user confirmation) by deleting all simulations, monads, samplings, and trials.

All the base inputs files will be kept, so previously run scripts should still work as expected. If the user aborts the reset, the user will then be asked if they want to continue with the script.

Keyword Arguments

force_reset::Bool: If true, skips the user confirmation prompt. Default is false.

force_continue::Bool: If true, skips the user confirmation prompt for continuing with the script after aborting the reset. Default is false.

pcvct.resetFolder — Method

resetFolder(location::Symbol, folder::String)

Reset a specific folder in the database for a given location, removing the variations database and variations folder.

Arguments

location: The location of the folder to reset.

folder: The name of the folder to reset.

pcvct.rm_hpc_safe — Method

rm_hpc_safe(path::String; force::Bool=false, recursive::Bool=false)

Remove files and take care if on an HPC since the NFS filesystem can leave behind extra files.

If on an HPC, move deleted files into the hidden directory data/.trash/ in a time-stamped folder.