Export

This file holds the functions for exporting a simulation to a user_project format.

pcvct.createExportFolder — Method

createExportFolder(simulation::Simulation, export_folder::AbstractString)

Create and populate the export folder for a simulation.

pcvct.exportIntracellular — Method

exportIntracellular(simulation::Simulation, export_folder::AbstractString)

Export the intracellular model for a simulation to the export folder.

pcvct.exportSimulation — Function

exportSimulation(simulation_id::Integer[, export_folder::AbstractString])

Create a user_project folder from a simulation that can be loaded into PhysiCell.

Warning: not all features in drbergman/PhysiCell/latest/release are not supported in MathCancer/PhysiCell.

Arguments

simulation_id::Integer: the id of the simulation to export. Can also be a Simulation object.

export_folder::AbstractString: the folder to export the simulation to. Default is the simulation output folder.

Returns

export_folder::AbstractString: the folder where the simulation was exported to

pcvct.revertConfig — Method

revertConfig(export_folder::AbstractString, physicell_version::AbstractString)

Revert the config folder in the export folder to the given PhysiCell version.

pcvct.revertCustomCPP — Method

revertCustomCPP(path_to_custom_modules::AbstractString, physicell_version::AbstractString)

Revert the custom cpp file in the export folder to the given PhysiCell version.

pcvct.revertCustomHeader — Method

revertCustomHeader(path_to_custom_modules::AbstractString, physicell_version::AbstractString)

Revert the custom header file in the export folder to the given PhysiCell version.

pcvct.revertCustomModules — Method

revertCustomModules(export_folder::AbstractString, physicell_version::AbstractString)

Revert the custom modules in the export folder to the given PhysiCell version.

pcvct.revertMain — Method

revertMain(export_folder::AbstractString, physicell_version::AbstractString)

Revert the main.cpp file in the export folder to the given PhysiCell version.

pcvct.revertMakefile — Method

revertMakefile(export_folder::AbstractString, physicell_version::AbstractString)

Revert the Makefile in the export folder to the given PhysiCell version.

pcvct.revertSimulationFolderToCurrentPhysiCell — Method

revertSimulationFolderToCurrentPhysiCell(export_folder::AbstractString, physicell_version::AbstractString)

Revert the simulation folder to the given PhysiCell version.

pcvct.setECMSetupElement — Method

setECMSetupElement(xml_doc::XMLDocument)

Set up the ECM element in the XML document to support the ECM module.