Globals

Global variables used in pcvct.

pcvct.PCVCTGlobals — Type

PCVCTGlobals

A mutable struct to hold global variables for the PCVCT package.

Fields

initialized::Bool: Indicates whether the project database has been initialized.

data_dir::String: The path to the data directory. This is set when the model manager is initialized.

physicell_dir::String: The path to the PhysiCell directory. This is set when the model manager is initialized.

inputs_dict::Dict{Symbol,Any}: A dictionary that maps the types of inputs to the data that defines how they are set up and what they can do. Read in from the inputs.toml file in the data directory.

project_locations::ProjectLocations: The global ProjectLocations object that contains information about the locations of input files in the project.

db::SQLite.DB: The database connection object to the central SQLite database for the open project. Set this with initializeModelManager.

strict_physicell_check::Bool: Indicates whether to perform strict checks on the PhysiCell directory for reproducibility. If true, requires a clean git folder (in particular, not downloaded) to skip recompile.

current_physicell_version_id::Int: The ID of the current version of PhysiCell being used as defined in the database. This is set when the model manager is initialized.

physicell_compiler::String: The compiler used to compile the PhysiCell code. This is set when the model manager is initialized.

march_flag::String: The march flag to be used when compiling the code. If running on an HPC, this is set to "x86-64" which will work across different CPU manufacturers that may be present on an HPC. Otherwise, set to "native".

run_on_hpc::Bool: A boolean that indicates whether the code is running on an HPC environment. This is set to true if the sbatch command is available when compiling pcvct.

sbatch_options::Dict{String,Any}: A dictionary that will be used to pass options to the sbatch command. The keys are the flag names and the values are the values used for the flag. This is initialized using defaultJobOptions and can be modified using setJobOptions.

max_number_of_parallel_simulations::Int: The maximum number of parallel simulations that can be run at once. If running on an HPC, this is ignored and instead pcvct will queue one job per simulation.

path_to_python::Union{Missing,String}: The path to the python executable for running PhysiCell Studio. See runStudio.

path_to_studio::Union{Missing,String}: The path to the PhysiCell Studio directory. See runStudio.

path_to_magick::Union{Missing,String}: The path to the ImageMagick installation. See makeMovie.

path_to_ffmpeg::Union{Missing,String}: The path to the FFmpeg installation. See makeMovie.

pcvct.centralDB — Method

centralDB()

Get the database global variable for the current project.

pcvct.dataDir — Method

dataDir()

Get the data directory global variable for the current project.

pcvct.inputsDict — Method

inputsDict()

Get the inputs dictionary global variable for the current project.

pcvct.physicellDir — Method

physicellDir()

Get the PhysiCell directory global variable for the current project.

pcvct.projectLocations — Method

projectLocations()

Get the project locations global variable for the current project.