Core

Core functionality for pcvct.

pcvct.PHYSICELL_CPP — Constant

PHYSICELL_CPP::String

The compiler used to compile the PhysiCell code. This is set when the model manager is initialized.

pcvct.current_physicell_version_id — Constant

current_physicell_version_id::Int

The ID of the current version of PhysiCell being used as defined in the database. This is set when the model manager is initialized.

pcvct.data_dir — Constant

data_dir::String

The path to the data directory. This is set when the model manager is initialized.

pcvct.initialized — Constant

initialized::Bool

A boolean that indicates whether a project database has been initialized to be used with pcvct.

pcvct.inputs_dict — Constant

inputs_dict::Dict{Symbol, Any}

A dictionary that maps the types of inputs to the data that defines how they are set up and what they can do. Read in from the inputs.toml file in the data directory.

pcvct.march_flag — Constant

march_flag::String

The march flag to be used when compiling the code.

If running on an HPC, this is set to "x86-64" which will work across different CPU manufacturers that may be present on an HPC. Otherwise, set to "native".

pcvct.max_number_of_parallel_simulations — Constant

max_number_of_parallel_simulations::Int

The maximum number of parallel simulations that can be run at once. If running on an HPC, this is ignored and instead pcvct will queue one job per simulation.

pcvct.physicell_dir — Constant

physicell_dir::String

The path to the PhysiCell directory. This is set when the model manager is initialized.

pcvct.run_on_hpc — Constant

run_on_hpc::Bool

A boolean that indicates whether the code is running on an HPC environment.

This is set to true if the sbatch command is available when compiling pcvct.

pcvct.sbatch_options — Constant

sbatch_options::Dict{String,Any}

A dictionary that will be used to pass options to the sbatch command.

The keys are the flag names and the values are the values used for the flag. This is initialized using defaultJobOptions and can be modified using setJobOptions.

pcvct.baseToExecutable — Function

baseToExecutable(s::String)

Convert a string to an executable name based on the operating system. If the operating system is Windows, append ".exe" to the string.

pcvct.constituentsType — Method

constituentsType(T::Type{<:AbstractTrial})
constituentsType(T::AbstractTrial)

Return the type of the constituents of T. Used in the readConstituentIDs function.

pcvct.constituentsTypeFilename — Method

constituentsTypeFilename(T::Type{<:AbstractTrial})
constituentsTypeFilename(T::AbstractTrial)

Return the filename of the constituents of T. Used in the readConstituentIDs function.

pcvct.getMonadIDs — Method

getMonadIDs()

Return a vector of all monad IDs in the database.

Alternate forms take a monad, sampling, or trial object (or an array of any combination of them) and return the corresponding monad IDs.

Examples

getMonadIDs() # all monad IDs in the database
getMonadIDs(monad) # just a vector with the monad ID, i.e. [monad.id]
getMonadIDs(sampling) # all monad IDs in a sampling
getMonadIDs(trial) # all monad IDs in a trial
getMonadIDs([trial1, trial2]) # all monad IDs between trial1 and trial2

pcvct.getSamplingSimulationIDs — Method

getSamplingSimulationIDs(sampling_id::Int)

Internal function to get the simulation IDs for a given sampling ID. Users should use getSimulationIDs instead.

pcvct.getSimulationIDs — Method

getSimulationIDs()

Return a vector of all simulation IDs in the database.

Alternate forms take a simulation, monad, sampling, or trial object (or an array of any combination of them) and return the corresponding simulation IDs.

Examples

getSimulationIDs() # all simulation IDs in the database
getSimulationIDs(simulation) # just a vector with the simulation ID, i.e. [simulation.id]
getSimulationIDs(monad) # all simulation IDs in a monad
getSimulationIDs(sampling) # all simulation IDs in a sampling
getSimulationIDs(trial) # all simulation IDs in a trial
getSimulationIDs([trial1, trial2]) # all simulation IDs between trial1 and trial2

pcvct.getTrialMonads — Method

getTrialMonads(trial_id::Int)

Internal function to get the monad IDs for a given trial ID. Users should use getMonadIDs instead.

pcvct.getTrialSimulationIDs — Method

getTrialSimulationIDs(trial_id::Int)

Internal function to get the simulation IDs for a given trial ID. Users should use getSimulationIDs instead.

pcvct.initializeModelManager — Method

initializeModelManager(path_to_physicell::String, path_to_data::String)

Initialize the VCT environment by setting the paths to PhysiCell and data directories, and initializing the database.

Arguments

path_to_physicell::String: Path to the PhysiCell directory as either an absolute or relative path.

path_to_data::String: Path to the data directory as either an absolute or relative path.

pcvct.initializeModelManager — Method

initializeModelManager()

Initialize the VCT environment assuming that the PhysiCell and data directories are in the current working directory.

pcvct.lowerClassString — Method

lowerClassString(T::AbstractTrial)
lowerClassString(T::Type{<:AbstractTrial})

Return the lowercase string representation of the type of T, excluding the module name. Without this, it may return, e.g., Main.pcvct.Sampling.

Examples

lowerClassString(Simulation) # "simulation"
lowerClassString(Simulation(1)) # "simulation"

pcvct.pcvctLogo — Method

pcvctLogo()

Return a string representation of the awesome pcvct logo.

pcvct.readConstituentIDs — Method

readConstituentIDs(T::AbstractTrial)

Read a CSV file containing constituent IDs from T and return them as a vector of integers.

For a trial, this is the sampling IDs. For a sampling, this is the monad IDs. For a monad, this is the simulation IDs.

Examples

ids = readConstituentIDs(Sampling(1)) # read the IDs of the monads in sampling 1
ids = readConstituentIDs(Sampling, 1) # identical to above but does not need to create the Sampling object

ids = readConstituentIDs(Monad, 1) # read the IDs of the simulations in monad 1
ids = readConstituentIDs(Trial, 1) # read the IDs of the samplings in trial 1

pcvct.setMarchFlag — Method

setMarchFlag(flag::String)

Set the march flag to flag. Used for compiling the PhysiCell code.

pcvct.setNumberOfParallelSims — Method

setNumberOfParallelSims(n::Int)

Set the maximum number of parallel simulations to n.

pcvct.trialFolder — Method

trialFolder(T::Type{<:AbstractTrial}, id::Int)

Return the path to the folder for a given subtype of AbstractTrial and ID.

Examples

trialFolder(Simulation, 1)
# output
"abs/path/to/data/outputs/simulations/1"

pcvct.trialFolder — Method

trialFolder(T::Type{<:AbstractTrial})

Return the path to the folder for the AbstractTrial object, T.

Examples

simulation = Simulation(1)
trialFolder(Simulation)
# output
"abs/path/to/data/outputs/simulations/1"