PhysiCell Version

Manage the PhysiCell version used in the project.

pcvct.currentPhysiCellVersionID — Method

currentPhysiCellVersionID()

Get the current PhysiCell version ID.

pcvct.getCommitHashToTagDict — Method

getCommitHashToTagDict(dir::String)

Get a dictionary mapping commit hashes to tags in the git repository at dir.

pcvct.gitDirectoryIsClean — Method

gitDirectoryIsClean(dir::String)

Check if the git directory is clean (i.e., no uncommitted changes).

pcvct.gitRemotes — Method

gitRemotes(dir::String)

Get the remotes for the git repository at dir.

pcvct.physiCellCommitHash — Method

physicellCommitHash()

Get the commit hash for the current PhysiCell version.

pcvct.physicellInfo — Method

physicellInfo()

Return a string representing the PhysiCell version information to display on initializing the model manager.

pcvct.physicellIsGit — Method

physicellIsGit()

Check if the PhysiCell directory is a git repository.

pcvct.physicellVersion — Method

physicellVersion()
physiCellVersion(physicell_version_id::Int)
physiCellVersion(simulation::Simulation)

Get the PhysiCell version from the database or, if not in the database, from the VERSION.txt file.

pcvct.repoOwner — Method

repoOwner(commit_hash::String, tag::String)

Get the owner of the repository for a given commit hash and tag.

pcvct.resolvePhysiCellVersionID — Method

resolvePhysiCellVersionID()

Get the PhysiCell version ID from the database, adding it to the database if it doesn't exist.