Pruner

Prune files from a simulation immediately after finishing the simulation.

To motivate this functionality, consider the following scenario. A user has been testing their model, including making movies, and is ready to do a large virtual clinical trial with thousands of simulations. Saving all the SVGs will require gigabytes of storage, which is not ideal for the user. The user could choose to create a new variation on the SVG parameters (e.g., increase the SVG save interval), but then pcvct will not be able to reuse previous simulations as they have different variation IDs. Alternatively, the user can use the PruneOptions to delete the SVGs after each simulation is finished. This way, there are fewer variations in the database and more capability to reuse simulations.

pcvct.PruneOptions — Type

PruneOptions

Automatically prune some of the generated output files from a simulation.

Fields

prune_svg::Bool=false: Prune SVG files

prune_txt::Bool=false: Prune TXT files

prune_mat::Bool=false: Prune MAT files

prune_initial::Bool=false: If any of the above are true, also prune the initial files for that type

prune_final::Bool=false: If any of the above are true, also prune the final files for that type

Examples

julia> PruneOptions(prune_svg=true, prune_txt=true, prune_mat=true)
PruneOptions
  prune_mat: Bool true
  prune_svg: Bool true
  prune_txt: Bool true
  prune_xml: Bool false
  prune_initial: Bool false
  prune_final: Bool false

pcvct.pruneSimulationOutput — Function

pruneSimulationOutput(simulation_id::Integer, prune_options::PruneOptions=PruneOptions())

Prune the output files from a simulation.

Arguments

simulation_id::Integer: The ID of the PhysiCell simulation. A Simulation object can also be passed in.

prune_options::PruneOptions=PruneOptions(): The options for pruning the output files. See PruneOptions for more information.

```