PCVCT Upgrade

Functionality for upgrading the database to match the current version of pcvct.

pcvct.populateTableOnFeatureSubset — Method

populateTableOnFeatureSubset(db::SQLite.DB, source_table::String, target_table::String; column_mapping::Dict{String, String}=Dict{String,String}())

Populate a target table with data from a source table, using a column mapping if provided.

pcvct.upgradePCVCT — Method

upgradePCVCT(from_version::VersionNumber, to_version::VersionNumber, auto_upgrade::Bool)

Upgrade the PCVCT database from one version to another.

The upgrade process is done in steps, where each step corresponds to a milestone version. The function will apply all necessary upgrades until the target version is reached. If auto_upgrade is true, the function will automatically apply all upgrades without prompting. Otherwise, it will prompt the user for confirmation before large upgrades.

pcvct.upgradeToVX_Y_Z — Function

upgradeToX_Y_Z(auto_upgrade::Bool)

Upgrade the database to pcvct version X.Y.Z. Each milestone version has its own upgrade function.