VCTCompilation

Compile a PhysiCell project in pcvct. Includes the necessary compiler macros and checks PhysiCell version by the commit hash of the PhysiCell repository.

pcvct.compilerFlags — Method

compilerFlags(S::AbstractSampling)

Generate the compiler flags for the given sampling object S.

Generate the necessary compiler flags based on the system and the macros defined in the sampling object S. If the required macros differ from a previous compilation (as stored in macros.txt), then recompile.

Returns

cflags::String: The compiler flags as a string.

recompile::Bool: A boolean indicating whether recompilation is needed.

clean::Bool: A boolean indicating whether cleaning is needed.

pcvct.loadCustomCode — Method

loadCustomCode(S::AbstractSampling[; force_recompile::Bool=false])

Load and compile custom code for a simulation, monad, or sampling.

Determines if recompilation is necessary based on the previously used macros. If compilation is required, copy the PhysiCell directory to a temporary directory to avoid conflicts. Then, compile the project, recording the output and error in the custom_codes folder used. Move the compiled executable into the custom_codes folder and the temporary PhysiCell folder deleted.