VCTDeletion

This file holds functionality for deleting output.

pcvct.deleteSimulationsByStatus — Function

deleteSimulationsByStatus(status_codes_to_delete::Vector{String}=["Failed"]; user_check::Bool=true)

Delete simulations from the database based on their status codes.

# Arguments
- `status_codes_to_delete::Vector{String}`: A vector of status codes for which simulations should be deleted. Default is `["Failed"]`.
- `user_check::Bool`: If `true`, prompts the user for confirmation before deleting simulations. Default is `true`.

# Description
This function performs the following steps:
1. Queries the database to retrieve simulation IDs and their corresponding status codes.
2. Iterates over the provided `status_codes_to_delete`.
3. For each status code, retrieves the simulation IDs that match the status code.
4. If no simulations match the status code, it continues to the next status code.
5. If `user_check` is `true`, prompts the user for confirmation before deleting the simulations.
6. If the user confirms, deletes the simulations with the matching status code.

# Returns
- Nothing. The function performs database operations and deletes records as needed.

pcvct.eraseSimulationID — Method

eraseSimulationID(simulation_id::Int; monad_id::Union{Missing,Int}=missing, T::Union{Missing,AbstractTrial}=missing)

Erase a simulation ID from the monad(s) it belongs to. Used when a simulation fails to run and needs to be removed from the monad's `simulations.csv`.

# Arguments
- `simulation_id::Int`: The ID of the simulation to be erased.
- `monad_id::Union{Missing,Int}`: The ID of the monad associated with the simulation. If not provided, it will be inferred from the simulation ID. Default is `missing`.
- `T::Union{Missing,AbstractTrial}`: An optional parameter for additional context. Default is `missing`.

# Description
This function performs the following steps:
1. If `monad_id` is missing, it constructs a query to find the monad ID associated with the given `simulation_id`.
2. Retrieves the list of simulation IDs associated with the monad.
3. Finds the index of the given `simulation_id` in the list.
4. If the `simulation_id` is not found, the function returns early.
5. If the monad contains only the given `simulation_id`, it deletes the monad and any dependent records, but retains the simulation in the database for output file checks.
6. If the monad contains multiple simulations, it removes the given `simulation_id` from the list and updates the records.

# Returns
- Nothing. The function performs database operations and updates records as needed.