VCTDeletion

Safely delete output from a pcvct project.

pcvct.deleteSimulations — Method

deleteSimulations(simulation_ids::AbstractVector{<:Union{Integer,Missing}}; delete_supers::Bool=true, and_constraints::String="")

Deletes the simulations with the input IDs from the database and from the data/outputs/simulations folder.

Works with any vector of integers or a single integer. If delete_supers is true, it will also delete any monads, samplings, and trials that no longer have any simulations associated with them. It is recommended to leave this to true to keep the database clean. The and_constraints argument allows for additional SQLite conditions to be added to the WHERE clause of the SQLite query. Use this only after inspecting the simulations table in the data/vct.db database. Note: deleteSimulation is an alias for deleteSimulations.

Examples

deleteSimulations(1:3)
deleteSimulations(4)
deleteSimulations(1:100; and_constraints="AND config_id = 1") # delete simulations with IDs 1 to 100 that have config_id = 1

pcvct.deleteSimulationsByStatus — Function

deleteSimulationsByStatus(status_codes_to_delete::Vector{String}=["Failed"]; user_check::Bool=true)

Delete simulations from the database based on their status codes.

The list of possible status codes is: "Not Started", "Queued", "Running", "Completed", "Failed".

Arguments

status_codes_to_delete::Vector{String}: A vector of status codes for which simulations should be deleted. Default is ["Failed"].

user_check::Bool: If true, prompts the user for confirmation before deleting simulations. Default is true.

pcvct.eraseSimulationID — Method

eraseSimulationID(simulation_id::Int[; monad_id::Union{Missing,Int}=missing])

Erase a simulation ID from the simulations.csv file of the monad it belongs to.

If monad_id is not provided, the function will infer it from the simulation ID. If the monad contains only the given simulation ID, the monad will be deleted. This is used when running simulations if they error so that the monads no longer rely on them, but the simulation output can still be checked.

pcvct.resetDatabase — Method

resetDatabase()

Reset the database (after user confirmation) by deleting all simulations, monads, samplings, and trials.

All the base inputs files will be kept, so previously run scripts should still work as expected. If the user aborts the reset, the user will then be asked if they want to continue with the script.

Keyword Arguments

force_reset::Bool: If true, skips the user confirmation prompt. Default is false.

force_continue::Bool: If true, skips the user confirmation prompt for continuing with the script after aborting the reset. Default is false.