VCTLoader

Load PhysiCell data into useful forms for downstream analysis. This may be split off into its own module or even package eventually, likely with VCTAnalysis.jl.

pcvct.PhysiCellSequence — Type

PhysiCellSequence

A sequence of PhysiCell snapshots.

Fields

path_to_folder::String: The path to the folder containing the PhysiCell-defined simulation output.

snapshots::Vector{PhysiCellSnapshot}: A vector of PhysiCell snapshots.

cell_type_to_name_dict::Dict{Int, String}: A dictionary mapping cell type IDs to cell type names.

labels::Vector{String}: A vector of cell data labels.

substrate_names::Vector{String}: A vector of substrate names.

pcvct.PhysiCellSnapshot — Type

PhysiCellSnapshot

A single snapshot of a PhysiCell simulation.

The cells, substrates, and mesh fields may remain empty until they are needed for analysis.

Fields

path_to_folder::String: The path to the folder containing the PhysiCell simulation output.

index::Union{Int, Symbol}: The index of the snapshot. Can be an integer or a symbol (:initial or :final).

time::Float64: The time of the snapshot (in minutes).

cells::DataFrame: A DataFrame containing cell data.

substrates::DataFrame: A DataFrame containing substrate data.

mesh::Dict{String, Vector{Float64}}: A dictionary containing mesh data.

pcvct.getCellDataSequence — Method

getCellDataSequence(simulation_id::Integer, labels::Vector{String}; include_dead::Bool=false, include_cell_type::Bool=false)

Return a dictionary where the keys are cell IDs from the PhysiCell simulation and the values are NamedTuples containing the time and the values of the specified labels for that cell. For scalar values, such as volume, the values are in a length N vector, where N is the number of snapshots in the simulation. In the case of a label that has multiple columns, such as position, the values are concatenated into a length(snapshots) x number of columns array.

Arguments

simulation_id::Integer: The ID of the PhysiCell simulation. Alternatively, can be a Simulation object or a PhysiCellSequence object.

labels::Vector{String}: The labels to extract from the cell data. If a label has multiple columns, such as position, the columns are concatenated into a single array. Alternatively, a single label string can be passed.

include_dead::Bool=false: Whether to include dead cells in the data.

include_cell_type::Bool=false: Whether to include the cell type name in the data. Equivalent to including "cell_type_name" in labels.

Examples

data = getCellDataSequence(sequence, ["position", "elapsed_time_in_phase"]; include_dead=true, include_cell_type=true)
data[1] # the first cell's data
data[1].position # an Nx3 array of the cell's position over time
data[1].elapsed_time_in_phase # an Nx1 array of the cell's elapsed time in phase over time
data[1].cell_type_name # an Nx1 array of the cell type name of the first cell over time