VCTModule

Core functionality for pcvct.

pcvct.getSimulationIDs — Method

getSimulationIDs()

Return a vector of all simulation IDs in the database.

Alternate forms take a simulation, monad, sampling, or trial object (or an array of any combination of them) and return the corresponding simulation IDs.

Examples

getSimulationIDs() # all simulation IDs in the database
getSimulationIDs(simulation) # just a vector with the simulation ID, i.e. [simulation.id]
getSimulationIDs(monad) # all simulation IDs in a monad
getSimulationIDs(sampling) # all simulation IDs in a sampling
getSimulationIDs(trial) # all simulation IDs in a trial
getSimulationIDs([trial1, trial2]) # all simulation IDs between trial1 and trial2

pcvct.initializeModelManager — Method

initializeModelManager(path_to_physicell::String, path_to_data::String)

Initialize the VCT environment by setting the paths to PhysiCell and data directories, and initializing the database.

Arguments

path_to_physicell::String: Path to the PhysiCell directory as either an absolute or relative path.

path_to_data::String: Path to the data directory as either an absolute or relative path.

pcvct.pcvctLogo — Method

pcvctLogo()

Return a string representation of the awesome pcvct logo.

pcvct.setNumberOfParallelSims — Method

setNumberOfParallelSims(n::Int)

Set the maximum number of parallel simulations to n.