Configuration

Interface with the configuration file necessary for PhysiCell simulations.

Provide functionality for accessing and modifying elements in any XML, including the PhysiCell configuration file, XML rules file, combined intracellular XML file, XML IC cell file, and XML IC ECM file.

pcvct.project_locations — Constant

project_locations::ProjectLocations

The global ProjectLocations object that contains information about the locations of input files in the project.

pcvct.ProjectLocations — Type

ProjectLocations

A struct that contains information about the locations of input files in the project.

The global instance of this struct is project_locations and is created by reading the inputs.toml file in the data_dir directory. It is instantiated with the parseProjectInputsConfigurationFile function.

Fields

all::NTuple{L,Symbol}: A tuple of all locations in the project.

required::NTuple{M,Symbol}: A tuple of required locations in the project.

varied::NTuple{N,Symbol}: A tuple of varied locations in the project.

pcvct.apoptosisPath — Method

apoptosisPath(cell_definition::String, path_elements::Vararg{AbstractString})

Return the XML path to the apoptosis for the given cell definition (or deeper if more path elements are given).

pcvct.attackRatesPath — Method

attackRatesPath(cell_definition::String, path_elements::Vararg{AbstractString})

Return the XML path to the attack rates for the given cell definition (or deeper if more path elements are given).

The first optional path element identifies the name of the cell type to attack. This can be provided in either of the following formats:

pcvct.attackRatesPath("cd8", "cancer")

pcvct.attackRatesPath("cd8", "attack_rate:name:cancer")

pcvct.cellDefinitionPath — Method

cellDefinitionPath(cell_definition::String, path_elements::Vararg{AbstractString})

Return the XML path to the cell definition or deeper if more path elements are given.

pcvct.cellInteractionsPath — Method

cellInteractionsPath(cell_definition::String, path_elements::Vararg{AbstractString})

Return the XML path to the cell interactions for the given cell definition (or deeper if more path elements are given).

pcvct.cellParameterName — Method

cellParameterName(column_name::String)

Return the short name of the varied parameter associated with a cell definition for the given column name used in creating a DataFrame summary table.

pcvct.columnName — Method

columnName(xml_path)

Return the column name corresponding to the given XML path.

Works on a vector of strings, an XMLPath object, or a ElementaryVariation object. Inverse of columnNameToXMLPath.

pcvct.columnNameToXMLPath — Method

columnNameToXMLPath(column_name::String)

Return the XML path corresponding to the given column name.

Inverse of columnName.

pcvct.createInputsTOMLTemplate — Method

createInputsTOMLTemplate(path_to_toml::String)

Create a template TOML file for the inputs configuration at the specified path.

This is something users should not be changing. It is something in the codebase to hopefully facilitate extending this framework to other ABM frameworks.

pcvct.createXMLFile — Method

createXMLFile(location::Symbol, M::AbstractMonad)

Create XML file for the given location and variation_id in the given monad.

The file is placed in $(location)_variations and can be accessed from there to run the simulation(s).

pcvct.customDataPath — Method

customDataPath(cell_definition::String, path_elements::Vararg{AbstractString})

Return the XML path to the custom data for the given cell definition (or deeper if more path elements are given).

pcvct.cyclePath — Method

cyclePath(cell_definition::String, path_elements::Vararg{AbstractString})

Return the XML path to the cycle for the given cell definition (or deeper if more path elements are given).

pcvct.deathPath — Method

deathPath(cell_definition::String, path_elements::Vararg{AbstractString})

Return the XML path to the death for the given cell definition (or deeper if more path elements are given).

pcvct.folderIsVaried — Method

folderIsVaried(location::Symbol, folder::String)

Return true if the location folder allows for varying the input files, false otherwise.

pcvct.getChildByAttribute — Method

getChildByAttribute(parent_element::XMLElement, path_element_split::Vector{<:AbstractString})

Get the child element of parent_element that matches the given tag and attribute.

pcvct.getContent — Method

getContent(xml_doc::XMLDocument, xml_path::Vector{<:AbstractString}; required::Bool=true)

Get the content of the element in the XML document that matches the given path. See retrieveElement.

pcvct.icCellParameterName — Method

icCellParameterName(name::String)

Return the short name of the varied parameter associated with a cell patch for the given name used in creating a DataFrame summary table.

pcvct.icCellsPath — Method

icCellsPath(cell_definition::AbstractString, patch_type::AbstractString, patch_id path_elements::Vararg{AbstractString})

Return the XML path to the IC cell patch for the given cell definition, patch type, and patch ID. Optionally, add more path_elements to the path as extra arguments.

Example

julia> icCellsPath("default", "disc", 1)
3-element Vector{String}:
 "cell_patches:name:default"
 "patch_collection:type:disc"
 "patch:ID:1"

julia> icCellsPath("default", "disc", 1, "x0")
4-element Vector{String}:
 "cell_patches:name:default"
 "patch_collection:type:disc"
 "patch:ID:1"
 "x0"

pcvct.icECMParameterName — Method

icECMParameterName(name::String)

Return the short name of the varied parameter associated with a ECM patch for the given name used in creating a DataFrame summary table.

pcvct.icECMPath — Method

icECMPath(layer_id::Int, patch_type::AbstractString, patch_id, path_elements::Vararg{AbstractString})

Return the XML path to the IC ECM patch for the given layerid, patchtype, and patchid. Optionally, add more pathelements to the path as extra arguments.

Example

julia> icECMPath(2, "ellipse", 1)
3-element Vector{String}:
 "layer:ID:2"
 "patch_collection:type:ellipse"
 "patch:ID:1"

julia> icECMPath(2, "elliptical_disc", 1, "density")
4-element Vector{String}:
 "layer:ID:2"
 "patch_collection:type:elliptical_disc"
 "patch:ID:1"
 "density"

pcvct.locationIDName — Method

locationIDName(location)

Return the name of the ID column for the location (as either a String or Symbol).

Examples

julia> pcvct.locationIDName(:config)
"config_id"

pcvct.locationIDNames — Method

locationIDNames()

Return the names of the ID columns for all locations.

pcvct.locationPath — Function

locationPath(location::Symbol, folder=missing)

Return the path to the location folder in the inputs directory.

If folder is not specified, the path to the location folder is returned.

pcvct.locationPath — Method

locationPath(location::Symbol, S::AbstractSampling)

Return the path to the location folder in the inputs directory for the AbstractSampling object.

pcvct.locationPath — Method

locationPath(input_folder::InputFolder)

Return the path to the location folder in the inputs directory for the InputFolder object.

pcvct.locationVariationIDName — Method

locationVariationIDName(location)

Return the name of the variation ID column for the location (as either a String or Symbol).

Examples

julia> pcvct.locationVariationIDName(:config)
"config_variation_id"

pcvct.locationVariationIDNames — Method

locationVariationIDNames()

Return the names of the variation ID columns for all varied locations.

pcvct.makeXMLPath — Method

makeXMLPath(xml_doc::XMLDocument, xml_path::AbstractVector{<:AbstractString})

Create (if it does not exist) and return the XML element relative to the root of the given XML document.

Arguments

xml_doc::XMLDocument: The XML document to start from.

xml_path::AbstractVector{<:AbstractString}: The path to the XML element to create or retrieve. Can be a string representing a child of the root element.

pcvct.makeXMLPath — Method

makeXMLPath(current_element::XMLElement, xml_path::AbstractVector{<:AbstractString})

Create (if it does not exist) and return the XML element relative to the given XML element.

Similar functionality to the shell command mkdir -p, but for XML elements.

Arguments

current_element::XMLElement: The current XML element to start from.

xml_path::AbstractVector{<:AbstractString}: The path to the XML element to create or retrieve. Can be a string representing a child of the current element.

pcvct.motilityPath — Method

motilityPath(cell_definition::String, path_elements::Vararg{AbstractString})

Return the XML path to the motility for the given cell definition (or deeper if more path elements are given).

pcvct.necrosisPath — Method

necrosisPath(cell_definition::String, path_elements::Vararg{AbstractString})

Return the XML path to the necrosis for the given cell definition (or deeper if more path elements are given).

pcvct.parseProjectInputsConfigurationFile — Method

parseProjectInputsConfigurationFile()

Parse the inputs.toml file in the data_dir directory and create a global ProjectLocations object.

pcvct.pathToICCell — Method

pathToICCell(simulation::Simulation)

Return the path to the IC cell file for the given simulation, creating it if it needs to be generated from an XML file.

pcvct.pathToICECM — Method

pathToICECM(simulation::Simulation)

Return the path to the IC ECM file for the given simulation, creating it if it needs to be generated from an XML file.

pcvct.phenotypePath — Method

phenotypePath(cell_definition::String, path_elements::Vararg{AbstractString})

Return the XML path to the phenotype for the given cell definition (or deeper if more path elements are given).

pcvct.prepareBaseFile — Method

prepareBaseFile(input_folder::InputFolder)

Return the path to the base XML file for the given input folder.

pcvct.prepareBaseRulesetsCollectionFile — Method

prepareBaseRulesetsCollectionFile(input_folder::InputFolder)

Return the path to the base XML file for the given input folder.

pcvct.prepareVariedInputFolder — Method

prepareVariedInputFolder(location::Symbol, sampling::Sampling)

Create the XML file for each monad in the sampling for the given location.

pcvct.prepareVariedInputFolder — Method

prepareVariedInputFolder(location::Symbol, M::AbstractMonad)

Create the XML file for the location in the monad.

pcvct.retrieveElement — Method

retrieveElement(xml_doc::XMLDocument, xml_path::Vector{<:AbstractString}; required::Bool=true)

Retrieve the element in the XML document that matches the given path.

If required is true, an error is thrown if the element is not found. Otherwise, nothing is returned if the element is not found.

pcvct.retrieveElementError — Method

retrieveElementError(xml_path::Vector{<:AbstractString}, path_element::String)

Throw an error if the element defined by xml_path is not found in the XML document, including the path element that caused the error.

pcvct.ruleParameterName — Method

ruleParameterName(name::String)

Return the short name of the varied parameter associated with a ruleset for the given name used in creating a DataFrame summary table.

pcvct.rulePath — Method

rulePath(cell_definition::AbstractString, behavior::AbstractString, path_elements::Vararg{AbstractString})

Return the XML path to the rule for the given cell definition and behavior.

Optionally, add more path_elements to the path as extra arguments.

Example

julia> rulePath("T cell", "attack rate")
2-element Vector{String}:
 "behavior_ruleset:name:T cell"
 "behavior:name:attack rate"

julia> rulePath("cancer", "cycle entry", "increasing_signals", "signal:name:oxygen", "half_max")
5-element Vector{String}:
 "behavior_ruleset:name:cancer"
 "behavior:name:cycle entry"
 "increasing_signals"
 "signal:name:oxygen"
 "half_max"

pcvct.sanitizePathElement — Method

sanitizePathElement(path_elements::String)

Disallow certain path elements to prevent security issues.

pcvct.shortConfigVariationName — Method

shortConfigVariationName(name::String)

Return the short name of the varied parameter in the configuration file for the given name used in creating a DataFrame summary table.

pcvct.shortICCellVariationName — Method

shortICCellVariationName(name::String)

Return the short name of the varied parameter in the IC cell file for the given name used in creating a DataFrame summary table.

pcvct.shortICECMVariationName — Method

shortICECMVariationName(name::String)

Return the short name of the varied parameter in the ECM file for the given name used in creating a DataFrame summary table.

pcvct.shortIntracellularVariationName — Method

shortIntracellularVariationName(name::String)

Return the short name of the varied parameter in the intracellular file for the given name used in creating a DataFrame summary table.

pcvct.shortLocationVariationID — Method

shortLocationVariationID(fieldname)

Return the short location variation ID name used in creating a DataFrame summary table.

Arguments

fieldname: The field name to get the short location variation ID for. This can be a symbol or a string.

pcvct.shortLocationVariationID — Method

shortLocationVariationID(type::Type, fieldname::Union{String, Symbol})

Return (as a type) the short location variation ID name used in creating a DataFrame summary table.

pcvct.shortRulesetsVariationName — Method

shortRulesetsVariationName(name::String)

Return the short name of the varied parameter in the rulesets collection file for the given name used in creating a DataFrame summary table.

pcvct.shortVariationName — Method

shortVariationName(location::Symbol, name::String)

Return the short name of the varied parameter for the given location and name used in creating a DataFrame summary table.

pcvct.tableName — Method

tableName(location)

Return the name of the table for the location (as either a String or Symbol).

Examples

julia> pcvct.tableName(:config)
"configs"

pcvct.updateField — Method

updateField(xml_doc::XMLDocument, xml_path::Vector{<:AbstractString}, new_value::Union{Int,Real,String})

Update the content of the element in the XML document that matches the given path with the new value. See retrieveElement.

pcvct.userParameterPath — Method

userParameterPath(field_name::String)

Return the XML path to the user parameter for the given field name.

pcvct.variationsTableName — Method

variationsTableName(location)

Return the name of the variations table for the location (as either a String or Symbol).